Conflicts:
	src/dlgpreferences.cpp
	src/library/dao/trackdao.cpp
	src/mixxx.cpp
	src/mixxx.h

+++ src/track/beatutils.cpp
@@ -9,9 +9,9 @@
 #include <QString>
 #include <QList>
 #include <QMap>
-#include <math.h>
 
 #include "track/beatutils.h"
+#include "util/math.h"
 
 // we are generous and assume the global_BPM to be at most 0.05 BPM far away
 // from the correct one
@@ -44,7 +44,7 @@ void BeatUtils::printBeatStatistics(const QVector<double>& beats, int SampleRate
 
         qDebug() << "Beat" << i << "local BPM:" << local_bpm;
 
-        local_bpm = floorf(local_bpm * kHistogramDecimalScale + 0.5) / kHistogramDecimalScale;
+        local_bpm = floor(local_bpm * kHistogramDecimalScale + 0.5) / kHistogramDecimalScale;
         frequency[local_bpm] += 1;
     }
 
@@ -93,7 +93,7 @@ QList<double> BeatUtils::computeWindowedBpmsAndFrequencyHistogram(
         double localBpm = 60.0 * windowSize / time;
 
         // round BPM to have two decimal places
-        double roundedBpm = floorf(localBpm * kHistogramDecimalScale + 0.5) /
+        double roundedBpm = floor(localBpm * kHistogramDecimalScale + 0.5) /
                 kHistogramDecimalScale;
 
         // add to local BPM to list and increment frequency count
@@ -248,7 +248,7 @@ double BeatUtils::calculateBpm(const QVector<double>& beats, int SampleRate, int
          double time = (beat_end - beat_start)/SampleRate;
          double local_bpm = 60.0 * N / time;
          // round BPM to have two decimal places
-         local_bpm = floorf(local_bpm * kHistogramDecimalScale + 0.5) / kHistogramDecimalScale;
+         local_bpm = floor(local_bpm * kHistogramDecimalScale + 0.5) / kHistogramDecimalScale;
 
          //qDebug() << "Local BPM beat " << i << ": " << local_bpm;
          if (!foundFirstCorrectBeat &&
